requirements.txt
fastapi
uvicorn[standard]
python-odf
python-multipart
pydantic